Revert: keep LR1110/semtech port untouched
Per review feedback (LuDuda), the nRF70 Wi-Fi location PAL must be fully
self-contained and must not require any changes to the LR1110 port.
Revert the LR1110/semtech-side conditional compilation and the split of
sid_pal_gnss_wifi_impl.c into sid_pal_wifi_lr1110_impl.c +
lr1110_location_priv.h that were introduced to make room for the nRF70
coexistence path.

The nRF70-Wi-Fi + LR1110-GNSS-only coexistence combination is dropped
for this PR; nRF70 Wi-Fi location now only targets standalone nRF70 or
nRF70 + SX1262 builds, and the symbol collision with the sx126x_location
NOSUPPORT stub is handled entirely from subsys/nrf70_location (see next
commit), not from subsys/semtech.
