+++ subsys/semtech/CMakeLists.txt
@@ -67,11 +67,7 @@ target_compile_definitions(smtc_lbm INTERFACE
 )
 
 add_subdirectory_ifdef(CONFIG_SIDEWALK_SUBGHZ_RADIO_SX126X sx126x)
-
-if(CONFIG_SIDEWALK_SUBGHZ_RADIO_SX126X AND NOT CONFIG_SIDEWALK_WIFI_LOCATION_NRF70)
-	add_subdirectory(sx126x_location)
-endif()
-
+add_subdirectory_ifdef(CONFIG_SIDEWALK_SUBGHZ_RADIO_SX126X sx126x_location)
 add_subdirectory_ifdef(CONFIG_SIDEWALK_SUBGHZ_RADIO_LR1110 lr11xx)
 add_subdirectory_ifdef(CONFIG_SIDEWALK_SUBGHZ_RADIO_LR1110 lr1110)
 add_subdirectory_ifdef(CONFIG_SIDEWALK_SUBGHZ_RADIO_LR1110 lr1110_location)
@@ -82,13 +78,9 @@ if(CONFIG_SIDEWALK_SUBGHZ_RADIO_SX126X)
 	target_link_libraries(sidewalk_radio INTERFACE
 		"-Wl,--start-group"
 		sid_pal_radio_sx126x_impl
+		sid_pal_radio_sx126x_location_impl
 		"-Wl,--end-group"
 	)
-	if(NOT CONFIG_SIDEWALK_WIFI_LOCATION_NRF70)
-		target_link_libraries(sidewalk_radio INTERFACE
-			sid_pal_radio_sx126x_location_impl
-		)
-	endif()
 endif()
 
 if(CONFIG_SIDEWALK_SUBGHZ_RADIO_LR1110)
